Move API Gateway traits to their own package
