+++ smithy-aws-apigateway-openapi/src/main/java/software/amazon/smithy/aws/apigateway/openapi/AddAuthorizers.java
@@ -18,10 +18,10 @@ package software.amazon.smithy.aws.apigateway.openapi;
 import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
-import software.amazon.smithy.aws.traits.apigateway.AuthorizerDefinition;
-import software.amazon.smithy.aws.traits.apigateway.AuthorizerIndex;
-import software.amazon.smithy.aws.traits.apigateway.AuthorizerTrait;
-import software.amazon.smithy.aws.traits.apigateway.AuthorizersTrait;
+import software.amazon.smithy.aws.apigateway.traits.AuthorizerDefinition;
+import software.amazon.smithy.aws.apigateway.traits.AuthorizerIndex;
+import software.amazon.smithy.aws.apigateway.traits.AuthorizerTrait;
+import software.amazon.smithy.aws.apigateway.traits.AuthorizersTrait;
 import software.amazon.smithy.model.node.Node;
 import software.amazon.smithy.model.node.ObjectNode;
 import software.amazon.smithy.model.shapes.ServiceShape;
