+++ smithy-aws-apigateway-openapi/src/main/java/software/amazon/smithy/aws/apigateway/openapi/AddCorsPreflightIntegration.java
@@ -26,8 +26,8 @@ import java.util.Set;
 import java.util.TreeSet;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
-import software.amazon.smithy.aws.traits.apigateway.IntegrationResponse;
-import software.amazon.smithy.aws.traits.apigateway.MockIntegrationTrait;
+import software.amazon.smithy.aws.apigateway.traits.IntegrationResponse;
+import software.amazon.smithy.aws.apigateway.traits.MockIntegrationTrait;
 import software.amazon.smithy.jsonschema.Schema;
 import software.amazon.smithy.model.node.ObjectNode;
 import software.amazon.smithy.model.traits.CorsTrait;
