+++ smithy-aws-apigateway-openapi/src/main/java/software/amazon/smithy/aws/apigateway/openapi/AddIntegrations.java
@@ -20,9 +20,9 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.logging.Logger;
-import software.amazon.smithy.aws.traits.apigateway.IntegrationTrait;
-import software.amazon.smithy.aws.traits.apigateway.IntegrationTraitIndex;
-import software.amazon.smithy.aws.traits.apigateway.MockIntegrationTrait;
+import software.amazon.smithy.aws.apigateway.traits.IntegrationTrait;
+import software.amazon.smithy.aws.apigateway.traits.IntegrationTraitIndex;
+import software.amazon.smithy.aws.apigateway.traits.MockIntegrationTrait;
 import software.amazon.smithy.model.node.Node;
 import software.amazon.smithy.model.node.ObjectNode;
 import software.amazon.smithy.model.shapes.OperationShape;
