+++ smithy-aws-apigateway-traits/src/main/resources/META-INF/smithy/aws.apigateway.json
