+++ smithy-aws-apigateway-traits/src/test/java/software/amazon/smithy/aws/apigateway/traits/ApiKeySourceTraitTest.java
@@ -1,0 +1,23 @@
+package software.amazon.smithy.aws.apigateway.traits;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.instanceOf;
+
+import org.junit.jupiter.api.Test;
+import software.amazon.smithy.model.node.Node;
+import software.amazon.smithy.model.shapes.ShapeId;
+import software.amazon.smithy.model.traits.Trait;
+import software.amazon.smithy.model.traits.TraitFactory;
+
+public class ApiKeySourceTraitTest {
+    @Test
+    public void registersTrait() {
+        TraitFactory factory = TraitFactory.createServiceFactory();
+        ShapeId id = ShapeId.from("smithy.example#Foo");
+        Trait trait = factory.createTrait(ApiKeySourceTrait.ID, id, Node.from("HEADER")).get();
+
+        assertThat(trait, instanceOf(ApiKeySourceTrait.class));
+        assertThat(factory.createTrait(ApiKeySourceTrait.ID, id, trait.toNode()).get(), equalTo(trait));
+    }
+}
