+++ smithy-aws-apigateway-traits/src/test/resources/software/amazon/smithy/aws/apigateway/traits/effective-authorizers.smithy
