+++ smithy-aws-traits/build.gradle.kts
@@ -19,5 +19,4 @@ extra["moduleName"] = "software.amazon.smithy.aws.traits"
 
 dependencies {
     api(project(":smithy-model"))
-    api(project(":smithy-utils"))
 }
