+++ smithy-aws-traits/src/main/resources/META-INF/services/software.amazon.smithy.model.traits.TraitService
@@ -6,12 +6,6 @@ software.amazon.smithy.aws.traits.DataPlaneTrait$Provider
 software.amazon.smithy.aws.traits.DataTrait$Provider
 software.amazon.smithy.aws.traits.ServiceTrait$Provider
 software.amazon.smithy.aws.traits.auth.UnsignedPayloadTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.AuthorizerTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.AuthorizersTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.RequestValidatorTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.ApiKeySourceTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.IntegrationTrait$Provider
-software.amazon.smithy.aws.traits.apigateway.MockIntegrationTrait$Provider
 software.amazon.smithy.aws.traits.clientendpointdiscovery.ClientEndpointDiscoveryTrait$Provider
 software.amazon.smithy.aws.traits.clientendpointdiscovery.ClientEndpointDiscoveryIdTrait$Provider
 software.amazon.smithy.aws.traits.clientendpointdiscovery.ClientDiscoveredEndpointTrait$Provider
