+++ smithy-aws-traits/src/main/resources/META-INF/smithy/manifest
@@ -1,4 +1,3 @@
 aws.api.json
-aws.apigateway.json
 aws.auth.json
 aws.protocols.json
