--- smithy-aws-traits/src/test/java/software/amazon/smithy/aws/traits/apigateway/MockIntegrationTraitTest.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-package software.amazon.smithy.aws.traits.apigateway;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-
-import org.junit.jupiter.api.Test;
-
-public class MockIntegrationTraitTest {
-    @Test
-    public void loadsValidTrait() {
-        MockIntegrationTrait trait = MockIntegrationTrait.builder()
-                .passThroughBehavior("when_no_templates")
-                .putRequestParameter("x", "y")
-                .build();
-
-        assertThat(trait.toBuilder().build(), equalTo(trait));
-    }
-}
